+++ lib/detail_page.dart
@@ -83,8 +83,8 @@ class _DetailPageState extends State<DetailPage> {
     final updatedData = {
       'titulo': _tituloController.text,
       'posicao': double.tryParse(_posicaoController.text) ?? 0,
-      'progresso': int.tryParse(_progressoController.text) ?? 0,
-      'total': int.tryParse(_totalController.text) ?? 0,
+      'progresso': double.tryParse(_progressoController.text) ?? 0,
+      'total': double.tryParse(_totalController.text) ?? 0,
       'inicio': _inicioController.text,
       'fim': _fimController.text,
       'tipo': _tipo,
